requirements.txt
# ================================================================
# GenerativeAug_LightDetection_for_CTC
# Python >= 3.10,  CUDA >= 12.x
# ================================================================

# ── PyTorch (install first, match your CUDA version) ──────────
# See https://pytorch.org/get-started/locally/
torch>=2.0.0
torchvision>=0.15.0

# ── Core ──────────────────────────────────────────────────────
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
Pillow>=9.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
PyYAML>=6.0
opencv-python>=4.8.0

# ── Detection: YOLO ───────────────────────────────────────────
ultralytics>=8.0.0

# ── Detection: Detectron2 (optional) ──────────────────────────
# Install from source: python -m pip install 'git+https://github.com/facebookresearch/detectron2.git'

# ── Detection: auxiliary ──────────────────────────────────────
onnx>=1.16.0
onnxruntime-gpu>=1.17.0

# ── GAN module ────────────────────────────────────────────────
# (all core deps already listed above)
